+++ pymarkdown/inline/inline_text_block_helper.py
@@ -3,7 +3,7 @@ Module to help with the parsing of text inline elements.
 """
 
 import logging
-from typing import List, Optional, Tuple, cast
+from typing import List, Optional, Set, Tuple, cast
 
 from pymarkdown.container_blocks.parse_block_pass_properties import (
     ParseBlockPassProperties,
@@ -90,6 +90,7 @@ class InlineTextBlockHelper:
         start_index = 0
         inline_blocks: List[MarkdownToken] = []
         end_string: Optional[str] = ""
+        pragma_line_numbers : Set[int]= set(list(parser_properties.pragma_lines.keys()))
 
         source_text, split_para_space = (
             InlineTextBlockHelper.__process_inline_text_block_prepare(
@@ -140,6 +141,7 @@ class InlineTextBlockHelper:
                 tabified_text,
                 newlines_encountered,
                 is_in_table,
+                pragma_line_numbers,
                 para_space,
             )
             if adj_newlines:
@@ -187,6 +189,7 @@ class InlineTextBlockHelper:
         tabified_text: Optional[str],
         newlines_encountered: int,
         is_in_table: bool,
+        pragma_line_numbers : Set[int],
         para_space: Optional[str] = None,
     ) -> Tuple[
         int,
@@ -260,6 +263,7 @@ class InlineTextBlockHelper:
             para_owner,
             tabified_text,
             tabified_remaining_line,
+            pragma_line_numbers
         )
 
         (
@@ -325,6 +329,7 @@ class InlineTextBlockHelper:
             last_line_number,
             last_column_number,
             end_string,
+            pragma_line_numbers
         )
 
         return (
@@ -576,6 +581,7 @@ class InlineTextBlockHelper:
         para_owner: Optional[ParagraphMarkdownToken],
         tabified_text: Optional[str],
         tabified_remaining_line: Optional[str],
+        pragma_line_numbers : Set[int],
     ) -> Tuple[
         InlineResponse,
         int,
@@ -606,6 +612,7 @@ class InlineTextBlockHelper:
                 line_number,
                 column_number,
                 coalesced_stack,
+                pragma_line_numbers
             )
             if inline_response.reduce_remaining_line_by:
                 (
@@ -811,6 +818,7 @@ class InlineTextBlockHelper:
         last_line_number: int,
         last_column_number: int,
         end_string: Optional[str],
+        pragma_line_numbers : Set[int]
     ) -> Tuple[
         int,
         int,
@@ -836,6 +844,7 @@ class InlineTextBlockHelper:
             remaining_line,
             did_line_number_change,
             was_column_number_reset,
+            pragma_line_numbers
         )
 
         (
@@ -905,6 +914,7 @@ class InlineTextBlockHelper:
         remaining_line: str,
         did_line_number_change: bool,
         was_column_number_reset: bool,
+        pragma_line_numbers : Set[int]
     ) -> Tuple[int, int, Optional[List[str]]]:
         if was_new_line:
             column_number = 1
@@ -926,6 +936,8 @@ class InlineTextBlockHelper:
                 column_number += selected_split_length
 
             line_number += 1
+            while line_number in pragma_line_numbers:
+                line_number += 1
             assert (
                 split_para_space
             ), "If we had a new line, we must have paragraph/setext text to handle."