+++ pymarkdown/plugins/rule_md_034.py
@@ -73,7 +73,7 @@ class RuleMd034(RulePlugin):
                 self.report_next_token_error(
                     context,
                     token,
-                    line_number_delta=line_number_delta,
+                    line_number_delta=line_number_delta + context.calc_pragma_offset(token, line_number_delta),
                     column_number_delta=column_number_delta,
                 )